refactor univ2 for function overriding

+++ contracts/uni-v2/utils/ERC20.sol
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: GPL-3.0
 pragma solidity >=0.8.0;
 
-import '../UniswapV2ERC20.sol';
+import "../UniswapV2ERC20.sol";
 
 contract ERC20 is UniswapV2ERC20 {
-    constructor(uint _totalSupply) {
+    constructor(uint256 _totalSupply) {
         _mint(msg.sender, _totalSupply);
     }
 }
